Remove uap specific RemoteExecutor code
