+++ src/CoreFx.Private.TestUtilities/src/CoreFx.Private.TestUtilities.csproj
@@ -19,7 +19,6 @@
     <Compile Include="System\TestEnvironment.cs" />
     <Compile Condition="'$(TargetGroup)' == 'netcoreapp' OR '$(TargetGroup)' == 'netcoreapp2.0'" Include="System\Diagnostics\RemoteExecutorTestBase.netcore.cs" />
     <Compile Condition="'$(TargetGroup)' == 'netfx'" Include="System\Diagnostics\RemoteExecutorTestBase.netfx.cs" />
-    <Compile Condition="'$(TargetGroup)' != 'uap'" Include="System\Diagnostics\RemoteExecutorTestBase.Process.cs" />
     <Compile Condition="'$(TargetGroup)' == 'uap'" Include="System\Diagnostics\RemoteExecutorTestBase.uap.cs" />
     <Compile Condition="'$(TargetsAOT)'=='true'" Include="System\Diagnostics\RemoteExecutorTestBase.aot.cs" />
     <Compile Condition="'$(TargetGroup)' == 'netfx'" Include="System\PlatformDetection.NetFx.cs" />
