+++ src/CoreFx.Private.TestUtilities/src/System/Diagnostics/RemoteExecutorTestBase.aot.cs
@@ -3,16 +3,16 @@
 // See the LICENSE file in the project root for more information.
 
 using System.IO;
+using System.Runtime.InteropServices;
 
 namespace System.Diagnostics
 {
     /// <summary>Base class used for all tests that need to spawn a remote process.</summary>
     public abstract partial class RemoteExecutorTestBase : FileCleanupTestBase
     {
-        protected static readonly string HostRunnerName = "xunit.console.netcore.exe";
-        protected  static readonly string HostRunner = TestConsoleApp;
+        protected static readonly string HostRunnerName = PlatformDetection.IsWindows ? "dotnet.exe" : "dotnet";
+        protected static readonly string HostRunner = Process.GetCurrentProcess().MainModule.FileName;
 
-        private static readonly string ExtraParameter = "";
+        private static readonly string ExtraParameter = TestConsoleApp;
     }
-
 }
