+++ src/CoreFx.Private.TestUtilities/src/System/Diagnostics/RemoteExecutorTestBase.cs
@@ -228,6 +228,70 @@ namespace System.Diagnostics
             return RemoteInvoke(GetMethodInfo(method), new[] { unparsedArg }, options, pasteArguments: false);
         }
 
+        /// <summary>Invokes the method from this assembly in another process using the specified arguments.</summary>
+        /// <param name="method">The method to invoke.</param>
+        /// <param name="args">The arguments to pass to the method.</param>
+        /// <param name="start">true if this function should Start the Process; false if that responsibility is left up to the caller.</param>
+        /// <param name="psi">The ProcessStartInfo to use, or null for a default.</param>
+        /// <param name="pasteArguments">true if this function should paste the arguments (e.g. surrounding with quotes); false if that responsibility is left up to the caller.</param>
+        private static RemoteInvokeHandle RemoteInvoke(MethodInfo method, string[] args, RemoteInvokeOptions options, bool pasteArguments = true)
+        {
+            options = options ?? new RemoteInvokeOptions();
+
+            // Verify the specified method returns an int (the exit code) or nothing,
+            // and that if it accepts any arguments, they're all strings.
+            Assert.True(method.ReturnType == typeof(void) || method.ReturnType == typeof(int) || method.ReturnType == typeof(Task<int>));
+            Assert.All(method.GetParameters(), pi => Assert.Equal(typeof(string), pi.ParameterType));
+
+            // And make sure it's in this assembly.  This isn't critical, but it helps with deployment to know
+            // that the method to invoke is available because we're already running in this assembly.
+            Type t = method.DeclaringType;
+            Assembly a = t.GetTypeInfo().Assembly;
+
+            // Start the other process and return a wrapper for it to handle its lifetime and exit checking.
+            ProcessStartInfo psi = options.StartInfo;
+            psi.UseShellExecute = false;
+
+            if (!options.EnableProfiling)
+            {
+                // Profilers / code coverage tools doing coverage of the test process set environment
+                // variables to tell the targeted process what profiler to load.  We don't want the child process 
+                // to be profiled / have code coverage, so we remove these environment variables for that process 
+                // before it's started.
+                psi.Environment.Remove("Cor_Profiler");
+                psi.Environment.Remove("Cor_Enable_Profiling");
+                psi.Environment.Remove("CoreClr_Profiler");
+                psi.Environment.Remove("CoreClr_Enable_Profiling");
+            }
+
+            // If we need the host (if it exists), use it, otherwise target the console app directly.
+            string metadataArgs = PasteArguments.Paste(new string[] { a.FullName, t.FullName, method.Name, options.ExceptionFile }, pasteFirstArgumentUsingArgV0Rules: false);
+            string passedArgs = pasteArguments ? PasteArguments.Paste(args, pasteFirstArgumentUsingArgV0Rules: false) : string.Join(" ", args);
+            string testConsoleAppArgs = ExtraParameter + " " + metadataArgs + " " + passedArgs;
+
+            // Uap console app is globally registered.
+            if (!File.Exists(HostRunner) && !PlatformDetection.IsUap)
+                throw new IOException($"{HostRunner} test app isn't present in the test runtime directory.");
+
+            if (options.RunAsSudo)
+            {
+                psi.FileName = "sudo";
+                psi.Arguments = HostRunner + " " + testConsoleAppArgs;
+            }
+            else
+            {
+                psi.FileName = HostRunner;
+                psi.Arguments = testConsoleAppArgs;
+            }
+
+            // Return the handle to the process, which may or not be started
+            return new RemoteInvokeHandle(options.Start ?
+                Process.Start(psi) :
+                new Process() { StartInfo = psi }, options,
+                a.FullName, t.FullName, method.Name
+                );
+        }
+
         private static MethodInfo GetMethodInfo(Delegate d)
         {
             // RemoteInvoke doesn't support marshaling state on classes associated with
@@ -262,25 +326,12 @@ namespace System.Diagnostics
                 MethodName = methodName;
             }
 
-            private int _exitCode;
             public int ExitCode
             {
                 get
                 {
-                    if (!PlatformDetection.IsUap)
-                    {
-                        Process.WaitForExit();
-                        return Process.ExitCode;
-                    }
-                    return _exitCode;
-                }
-                internal set
-                {
-                    if (!PlatformDetection.IsUap)
-                    {
-                        throw new PlatformNotSupportedException("ExitCode property can only be set in UWP");
-                    }
-                    _exitCode = value;
+                    Process.WaitForExit();
+                    return Process.ExitCode;
                 }
             }
             public Process Process { get; set; }
